Restructure a bit
