+++ src/simulation.rs
@@ -17,7 +17,6 @@ use investment::perform_agent_investment;
 pub mod prices;
 pub use prices::CommodityPrices;
 pub mod investment_tools;
-pub mod lcox;
 
 /// Run the simulation.
 ///
